refactor(components): merge useSearchLocation \& useChangeFocus
МЕТА ДАННЫЕ: refactor 12
